requirements.txt
absl-py==0.8.1
apturl==0.5.2
asn1crypto==0.24.0
astor==0.8.0
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.1
bleach==3.1.0
blis==0.4.1
boto3==1.10.16
botocore==1.13.16
Bottleneck==1.3.0
Brlapi==0.6.6
certifi==2019.9.11
chardet==3.0.4
Click==7.0
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
cymem==2.0.2
dataclasses==0.7
decorator==4.4.1
defer==1.0.6
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
docutils==0.15.2
en-core-web-sm==2.2.0
entrypoints==0.3
fastai==1.0.59
fastprogress==0.1.21
Flask==1.1.1
Flask-Cors==3.0.8
gast==0.3.2
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
httplib2==0.9.2
idna==2.8
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.0
json5==0.8.5
jsonschema==3.1.1
jupyter-client==5.3.4
jupyter-core==4.6.1
jupyterlab==1.2.2
jupyterlab-server==1.0.6
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-self-attention==0.42.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.1.1
MarkupSafe==1.0
matplotlib==3.1.1
mistune==0.8.4
mock==3.0.5
more-itertools==7.2.0
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==4.4.0
netifaces==0.10.4
notebook==6.0.2
numexpr==2.7.0
numpy==1.17.4
nvidia-ml-py3==7.352.0
oauth==1.0.1
olefile==0.45.1
opt-einsum==3.1.0
packaging==19.2
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
pdf2image==1.10.0
pdfminer.six==20191110
pexpect==4.2.1
pickleshare==0.7.5
Pillow==5.1.0
plac==1.1.3
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.0
ptyprocess==0.6.0
pycairo==1.16.2
pycrypto==2.6.1
pycryptodome==3.9.2
pycups==1.9.73
Pygments==2.4.2
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyocr==0.7.2
pyparsing==2.4.5
pyRFC3339==1.0
pyrsistent==0.15.5
pytesseract==0.3.0
python-apt==1.6.4
python-dateutil==2.6.1
python-debian==0.1.32
pytorch-transformers==1.0.0
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
pyzmq==18.1.0
regex==2019.11.1
reportlab==3.4.0
requests==2.22.0
requests-unixsocket==0.1.5
s3transfer==0.2.1
scikit-learn==0.21.3
scipy==1.3.2
SecretStorage==2.3.1
Send2Trash==1.5.0
sentencepiece==0.1.83
simplejson==3.13.2
six==1.13.0
sklearn==0.0
sortedcontainers==2.1.0
soupsieve==1.9.5
spacy==2.2.2
srsly==0.2.0
system-service==0.3
systemd-python==234
tb-nightly==1.14.0a20190301
tensorboard==1.15.0
tensorflow==1.14.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.4
tf-estimator-nightly==1.14.0.dev2019030115
thinc==7.3.1
torch==1.3.1
torchvision==0.4.2
tornado==6.0.3
tqdm==4.38.0
traitlets==4.3.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.7
usb-creator==0.3.3
wadllib==1.3.2
wasabi==0.4.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
wrapt==1.11.2
xkit==0.0.0
zipp==0.6.0
zope.interface==4.3.2
